Make LocationDetailsCard its own component.
